refactor: workspace split — koda-core (lib) + koda-cli (bin)
Phase 1 of the v0.2.0 client-server migration (#50).

Workspace structure:
- koda-core: library crate, zero terminal deps
  - Engine types (EngineEvent, EngineCommand, EngineSink)
  - Providers (Anthropic, Gemini, OpenAI-compat)
  - Tools (Bash, Edit, Write, AST, etc.)
  - MCP client, DB, config, approval logic
  - Inference loop with injectable callbacks
- koda-cli: binary crate, produces the 'koda' command
  - Terminal rendering (display, markdown, highlight)
  - Interactive approval (confirm, tui)
  - REPL (rustyline), onboarding, interrupt handling
  - CliSink: EngineSink impl for terminal

Key decouplings resolved:
- confirm::describe_action() → moved to koda_core::tools
- interrupt::is_interrupted() → injectable closure param
- SimpleSpinner → replaced with SpinnerStart/SpinnerStop sink events
- crate::app::create_provider() → koda_core::providers::create_provider()
- loop_guard::ask_continue_or_stop() → injectable LoopContinuation callback

293 tests pass (zero lost). Clippy clean. Zero behavior change.
