+++ Cargo.toml
@@ -1,104 +1,9 @@
-[package]
-name = "koda-agent"
-version = "0.1.4"
-edition = "2024"
-description = "A high-performance AI coding agent built in Rust"
-license = "MIT"
-repository = "https://github.com/lijunzh/koda"
-homepage = "https://github.com/lijunzh/koda"
-readme = "README.md"
-keywords = ["ai", "coding-agent", "llm", "cli", "rust"]
-categories = ["command-line-utilities", "development-tools"]
-authors = ["Lijun Zhu"]
-
-[[bin]]
-name = "koda"
-path = "src/main.rs"
-
-[dependencies]
-# Async runtime
-tokio = { version = "1", features = ["full"] }
-
-# CLI
-clap = { version = "4", features = ["derive", "env"] }
-
-
-# Database
-sqlx = { version = "0.8", features = ["runtime-tokio", "sqlite"] }
-
-# Serialization
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-toml = "0.8"
-
-# TUI / UX
-indicatif = "0.17"
-crossterm = "0.28"
-
-# Logging
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-tracing-appender = "0.2"
-
-# File system (respects .gitignore)
-ignore = "0.4"
-
-# Path normalization
-path-clean = "1"
-
-# Text search
-regex = "1"
-
-# File globbing
-glob = "0.3"
-
-# URL parsing (for SSRF protection)
-url = "2"
-
-# Unique IDs
-uuid = { version = "1", features = ["v4"] }
-
-# Base64 encoding (for image support)
-base64 = "0.22"
-
-# HTTP client (for LLM API)
-reqwest = { version = "0.12", features = ["json", "stream"] }
-
-# Streaming support
-futures-util = "0.3"
-
-# Syntax highlighting (same engine as bat/delta)
-syntect = { version = "5", default-features = false, features = ["default-syntaxes", "default-themes", "parsing", "html", "regex-fancy"] }
-once_cell = "1"
-
-# Error handling
-anyhow = "1"
-
-# Signal handling
-ctrlc = "3"
-
-# Async trait support
-async-trait = "0.1"
-
-# For reading user input
-rustyline = "15"
-
-# MCP (Model Context Protocol) client
-rmcp = { version = "0.16", features = ["client", "transport-child-process"] }
-tokio-util = { version = "0.7", features = ["codec"] }
-tree-sitter = "0.26.6"
-petgraph = "0.8.3"
-tree-sitter-rust = "0.24.0"
-tree-sitter-python = "0.25.0"
-tree-sitter-javascript = "0.25.0"
-tree-sitter-typescript = "0.23.2"
-
-[dev-dependencies]
-tempfile = "3"
-path-clean = "1"
+[workspace]
+members = ["koda-core", "koda-cli"]
+resolver = "3"
 
 [profile.release]
-strip = true       # Strip debug symbols (~3-4MB savings)
-lto = true          # Link-time optimization across all crates (~1-2MB savings)
-codegen-units = 1   # Single codegen unit for better optimization
-panic = "abort"     # Remove unwinding code (~200KB savings)
+strip = true
+lto = true
+codegen-units = 1
+panic = "abort"
