+++ koda-cli/Cargo.toml
@@ -1,0 +1,58 @@
+[package]
+name = "koda-cli"
+version = "0.1.4"
+edition = "2024"
+description = "A high-performance AI coding agent built in Rust"
+license = "MIT"
+repository = "https://github.com/lijunzh/koda"
+homepage = "https://github.com/lijunzh/koda"
+readme = "../README.md"
+keywords = ["ai", "coding-agent", "llm", "cli", "rust"]
+categories = ["command-line-utilities", "development-tools"]
+authors = ["Lijun Zhu"]
+
+[[bin]]
+name = "koda"
+path = "src/main.rs"
+
+[dependencies]
+# Core engine
+koda-core = { path = "../koda-core" }
+
+# CLI
+clap = { version = "4", features = ["derive", "env"] }
+
+# Async runtime
+tokio = { version = "1", features = ["full"] }
+
+# Serialization (needed for CLI-level config)
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+
+# TUI / Terminal
+crossterm = "0.28"
+
+# Syntax highlighting
+syntect = { version = "5", default-features = false, features = ["default-syntaxes", "default-themes", "parsing", "html", "regex-fancy"] }
+once_cell = "1"
+
+# Signal handling
+ctrlc = "3"
+
+# REPL
+rustyline = "15"
+
+# Logging (subscriber for CLI configuration)
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+tracing-appender = "0.2"
+
+# Error handling
+anyhow = "1"
+
+# Base64 encoding (for image input)
+base64 = "0.22"
+
+[dev-dependencies]
+tempfile = "3"
+
