+++ koda-core/src/db.rs
