+++ koda-core/src/memory.rs
