+++ koda-core/src/providers/openai_compat.rs
