+++ koda-core/src/tools/agent.rs
