+++ koda-core/src/tools/glob_tool.rs
