+++ koda-core/src/tools/memory.rs
