+++ koda-core/src/tools/mod.rs
@@ -290,3 +290,108 @@ mod tests {
         assert_eq!(result, PathBuf::from("/home/user/project"));
     }
 }
+
+// ── Tool action descriptions ──────────────────────────────────
+
+/// Generate a human-readable description of a tool action for approval prompts.
+pub fn describe_action(tool_name: &str, args: &serde_json::Value) -> String {
+    match tool_name {
+        "Bash" => {
+            let cmd = args
+                .get("command")
+                .or(args.get("cmd"))
+                .and_then(|v| v.as_str())
+                .unwrap_or("?");
+            format!("\x1b[1m{cmd}\x1b[0m")
+        }
+        "Delete" => {
+            let path = args
+                .get("file_path")
+                .or(args.get("path"))
+                .and_then(|v| v.as_str())
+                .unwrap_or("?");
+            let recursive = args
+                .get("recursive")
+                .and_then(|v| v.as_bool())
+                .unwrap_or(false);
+            if recursive {
+                format!("Delete directory (recursive): \x1b[1m{path}\x1b[0m")
+            } else {
+                format!("Delete: \x1b[1m{path}\x1b[0m")
+            }
+        }
+        "Write" => {
+            let path = args
+                .get("path")
+                .or(args.get("file_path"))
+                .and_then(|v| v.as_str())
+                .unwrap_or("?");
+            let overwrite = args
+                .get("overwrite")
+                .and_then(|v| v.as_bool())
+                .unwrap_or(false);
+            if overwrite {
+                format!("Overwrite file: \x1b[1m{path}\x1b[0m")
+            } else {
+                format!("Create file: \x1b[1m{path}\x1b[0m")
+            }
+        }
+        "Edit" => {
+            let path = if let Some(payload) = args.get("payload") {
+                payload
+                    .get("file_path")
+                    .or(payload.get("path"))
+                    .and_then(|v| v.as_str())
+                    .unwrap_or("?")
+            } else {
+                args.get("file_path")
+                    .or(args.get("path"))
+                    .and_then(|v| v.as_str())
+                    .unwrap_or("?")
+            };
+            format!("Edit file: \x1b[1m{path}\x1b[0m")
+        }
+        "WebFetch" => {
+            let url = args.get("url").and_then(|v| v.as_str()).unwrap_or("?");
+            format!("Fetch URL: \x1b[1m{url}\x1b[0m")
+        }
+        _ => format!("Execute: {tool_name}"),
+    }
+}
+
+#[cfg(test)]
+mod describe_action_tests {
+    use super::*;
+    use serde_json::json;
+
+    #[test]
+    fn test_describe_bash() {
+        let desc = describe_action("Bash", &json!({"command": "cargo build"}));
+        assert!(desc.contains("cargo build"));
+    }
+
+    #[test]
+    fn test_describe_delete() {
+        let desc = describe_action("Delete", &json!({"file_path": "old.rs"}));
+        assert!(desc.contains("old.rs"));
+    }
+
+    #[test]
+    fn test_describe_edit() {
+        let desc = describe_action("Edit", &json!({"payload": {"file_path": "src/main.rs"}}));
+        assert!(desc.contains("src/main.rs"));
+    }
+
+    #[test]
+    fn test_describe_write() {
+        let desc = describe_action("Write", &json!({"path": "new.rs"}));
+        assert!(desc.contains("Create file"));
+        assert!(desc.contains("new.rs"));
+    }
+
+    #[test]
+    fn test_describe_write_overwrite() {
+        let desc = describe_action("Write", &json!({"path": "x.rs", "overwrite": true}));
+        assert!(desc.contains("Overwrite"));
+    }
+}
