+++ koda-core/src/tools/reasoning.rs
